% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gldp_resource.R
\name{add_gldp_resource}
\alias{add_gldp_resource}
\title{Add a GeoLocator Data Resource}
\usage{
add_gldp_resource(
  pkg,
  resource_name,
  data,
  cast_type = FALSE,
  replace = FALSE,
  delim = ","
)
}
\arguments{
\item{pkg}{A GeoLocator Data Package object to which the resource will be added.}

\item{resource_name}{A character string specifying the name of the resource. This name is used
to locate the schema file.}

\item{data}{A data frame containing the data to be added as a resource. The data frame will be
adjusted according to the schema.}

\item{cast_type}{A logical value indicating whether the data frame should be cast to the types
specified in the schema. Defaults to \code{FALSE}.}

\item{replace}{If \code{TRUE}, allows an existing resource of the same name to be
replaced.}

\item{delim}{Delimiter for the CSV file(s) referenced in \code{data} (e.g. \verb{\\t}
for a tab-separated file).
Will be set as \code{delimiter} in the resource Table Dialect, so read functions
know how to read the file(s).
Ignored if \code{data} is a data frame.}
}
\value{
The updated GeoLocator Data Package object with the new resource added.
}
\description{
Add a standard GeoLocator-DP resource to a package. This essentially wraps
\code{\link[frictionless:add_resource]{frictionless::add_resource()}} but with prepartions steps of the data:
\itemize{
\item extending the schema with extra columns present in \code{data} when allowed by
\code{fieldsMatch};
\item adding missing schema columns to \code{data} with typed \code{NA} values when
needed;
\item reordering \code{data} columns to match the schema, because
\code{\link[frictionless:add_resource]{frictionless::add_resource()}} expects fields in schema order.
\item casting columns to schema types when \code{cast_type = TRUE}, for example
converting strings to dates, numbers, or integers when required by the
schema, so the stored resource is consistent with the declared schema.
}

For common package resources, the resource accessors are usually a more
convenient user-facing interface.
}
\seealso{
\code{\link[=tags]{tags()}}, \code{\link[=observations]{observations()}}, and the other resource accessors for the
usual user-facing way to modify standard GeoLocator-DP resources.
}
