% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_gldp.R
\name{ask_gldp}
\alias{ask_gldp}
\title{Interactive GeoLocator Data Package Creator}
\usage{
ask_gldp()
}
\value{
A Geolocator Data Package object.
}
\description{
This function interactively asks the user questions in the console
to collect all the information needed to create a GeoLocator Data Package.
It is a wrapper around \code{\link[=create_gldp]{create_gldp()}}.
}
