% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gldp.R
\name{check_gldp_pkg}
\alias{check_gldp_pkg}
\title{Validate the GeoLocator Data Package Metadata}
\usage{
check_gldp_pkg(pkg)
}
\arguments{
\item{pkg}{An object of class \code{"geolocatordp"} to be validated.}
}
\value{
The validated GeoLocator Data Package object (invisible).
}
\description{
This function checks that the provided object is a valid GeoLocator Data Package and conforms to
the expected structure. It verifies that the object has the \code{"geolocatordp"} class and performs
a general package validation using the \code{frictionless} package.
}
