% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecations.R
\name{config2tibble}
\alias{config2tibble}
\title{Deprecated: \code{config2tibble()}}
\usage{
config2tibble(
  file = Sys.getenv("R_CONFIG_FILE", "config.yml"),
  filter_return = TRUE
)
}
\arguments{
\item{file}{Character string specifying the path to the \code{config.yml} file.}

\item{filter_return}{Logical. If \code{TRUE}, only columns that vary across tags are returned.}
}
\value{
Same as \code{\link[=config_to_tibble]{config_to_tibble()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{config2tibble()} was renamed to \code{\link[=config_to_tibble]{config_to_tibble()}}.
}
\seealso{
\code{\link[=config_to_tibble]{config_to_tibble()}}.
}
