% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gldp.R
\name{create_gldp}
\alias{create_gldp}
\title{Create a GeoLocator Data Package}
\usage{
create_gldp(version = NULL, ...)
}
\arguments{
\item{version}{(optional) Version of GeoLocator Data Package to use. Defaults to the
GeoLocatoR default (latest).}

\item{...}{Additional descriptor fields to include}
}
\value{
A GeoLocator Data Package object.
}
\description{
Create a GeoLocator Data Package shell with a minimal descriptor.
}
\examples{
pkg <- create_gldp()

# Previous version
pkg <- create_gldp(version = "v0.2")

}
