% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecations.R
\name{create_gldp_geopressuretemplate}
\alias{create_gldp_geopressuretemplate}
\title{Deprecated: \code{create_gldp_geopressuretemplate()}}
\usage{
create_gldp_geopressuretemplate(directory = ".")
}
\arguments{
\item{directory}{A GeoPressureTemplate directory.}
}
\value{
A \code{geolocatordp} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{create_gldp_geopressuretemplate()} was superseded by \code{\link[=read_geopressuretemplate]{read_geopressuretemplate()}}.
This compatibility helper initializes \code{pkg = create_gldp()} and forwards the call.
}
