% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{edges}
\alias{edges}
\alias{edges<-}
\title{Get or set edges}
\usage{
edges(x)

edges(x) <- value
}
\arguments{
\item{x}{GeoLocator Data Package object, as returned by \code{read_geolocatordp()}.}

\item{value}{A data frame to assign as edges}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}} data frame with edges
}
\description{
\code{edges()} gets the edges from a GeoLocator Data Package object.

\verb{edges<-()} is the assignment equivalent.
It should only be used within other functions, where the expected data
structure can be guaranteed.
}
