% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gldp2zenodoRecord.R
\name{gldp2zenodoRecord}
\alias{gldp2zenodoRecord}
\title{Create or update a Zenodo Record based on the metadata from a GeoLocatoR Data Package}
\usage{
gldp2zenodoRecord(
  pkg,
  zenodo_record = zen4R::ZenodoRecord$new(),
  token = keyring::key_get(service = "ZENODO_PAT")
)
}
\arguments{
\item{pkg}{A Geolocator Data Package object}

\item{zenodo_record}{A \code{zen4R} ZenodoRecord object from which to extract metadata}

\item{token}{A Zenodo API token. Best practice is to store it in your keyring (see description
above), in which case the default value will retrieve it using \code{keyring::key_get()}.}
}
\value{
A ZenodoRecord object updated with the metadata from \code{pkg}.
}
\description{
This function update a \code{zen4R} ZenodoRecord based on the GeoLocator Data Package provided.
If no \code{zenodo_record} is provided, the function will create a new one.

Note that (1) the record is not deposited online (\code{zenodo_record} is a variable in your computer)
and (2) no data is actually retrieved/copied.
}
