% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecations.R
\name{gldp_to_zenodo}
\alias{gldp_to_zenodo}
\title{Deprecated: \code{gldp_to_zenodo()}}
\usage{
gldp_to_zenodo(...)
}
\arguments{
\item{...}{Unused.}
}
\description{
\code{gldp_to_zenodo()} was removed with no replacement.
}
