% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{measurements}
\alias{measurements}
\alias{measurements<-}
\title{Get or set measurements}
\usage{
measurements(x)

measurements(x) <- value
}
\arguments{
\item{x}{GeoLocator Data Package object, as returned by \code{read_geolocatordp()}.}

\item{value}{A data frame to assign as measurements}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}} data frame with measurements
}
\description{
\code{measurements()} gets the measurements from a GeoLocator Data Package object.

\verb{measurements<-()} is the assignment equivalent.
It should only be used within other functions, where the expected data
structure can be guaranteed.
}
