% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{observations}
\alias{observations}
\alias{observations<-}
\title{Get or set observations}
\usage{
observations(x)

observations(x) <- value
}
\arguments{
\item{x}{GeoLocator Data Package object, as returned by \code{read_geolocatordp()}.}

\item{value}{A data frame to assign as observations}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}} data frame with observations
}
\description{
\code{observations()} gets the observations from a GeoLocator Data Package object.

\verb{observations<-()} is the assignment equivalent.
It should only be used within other functions, where the expected data
structure can be guaranteed.
}
