% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params2o.R
\name{params2o}
\alias{params2o}
\title{Generate \code{observations} from GeoPressureR parameters}
\usage{
params2o(params)
}
\arguments{
\item{params}{A list of GeoPressureR parameter objects. These parameters should have been
generated during the GeoPressure worflow. See \href{https://raphaelnussbaumer.com/GeoPressureR/reference/param_create.html}{\code{GeoPressureR::param_create()} } for more information.}
}
\value{
A tibble with columns \code{tag_id}, \code{manufacturer}, \code{scientific_name}, \code{ring_number},
\code{model}, and \code{firwmare} (if \code{soi_settings} is present).
}
\description{
This function converts a list of parameters into a single \code{observations} tibble according to the
\href{https://raphaelnussbaumer.com/GeoLocator-DP/core/observations/}{GeoLocator Data Package specification }.
}
