% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params2t.R
\name{params2t}
\alias{params2t}
\title{Generate \code{tags} from GeoPressureR parameters}
\usage{
params2t(params)
}
\arguments{
\item{params}{A list of GeoPressureR parameter objects. These parameters should have been
generated during the GeoPressure worflow. See \href{https://raphaelnussbaumer.com/GeoPressureR/reference/param_create.html}{\code{GeoPressureR::param_create()} } for more information.}
}
\value{
A tibble with columns \code{tag_id}, \code{manufacturer}, \code{scientific_name}, \code{ring_number},
\code{model}, and \code{firmware} (if \code{soi_settings} is present).
}
\description{
This function converts a list of parameters into a single \code{tags} tibble according to the
\href{https://raphaelnussbaumer.com/GeoLocator-DP/core/tags/}{GeoLocator Data Package specification}.
}
