% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_to_observations.R
\name{params_to_observations}
\alias{params_to_observations}
\title{Generate \code{observations} from GeoPressureR parameters}
\usage{
params_to_observations(params)
}
\arguments{
\item{params}{A list of GeoPressureR parameter objects. These parameters should have been
generated during the GeoPressure workflow. See \href{https://geopressure.org/GeoPressureR/reference/param_create.html}{\code{GeoPressureR::param_create()} } for more information.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one equipment and one retrieval observation
per parameter object.
}
\description{
Convert a list of GeoPressureR parameter objects into a minimal
\code{observations} tibble following the GeoLocator Data Package structure.

One \code{equipment} row and one \code{retrieval} row are generated per parameter
object. When available, crop dates are used as observation datetimes and
known stationary locations are used to populate the first and last locations.
}
\seealso{
\code{\link[=params_to_tags]{params_to_tags()}} and \code{\link[=tags_to_measurements]{tags_to_measurements()}} for related
conversions from GeoPressureR objects to GeoLocator-DP resources.
}
