% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{paths}
\alias{paths}
\alias{paths<-}
\title{Get or set paths}
\usage{
paths(x)

paths(x) <- value
}
\arguments{
\item{x}{A GeoLocator Data Package object, as returned by \code{read_gldp()}.}

\item{value}{A data frame to assign as paths. Must conform to the paths schema specification.}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}} data frame with paths
}
\description{
\code{paths()} gets the paths from a GeoLocator Data Package object.

\verb{paths<-()} is the assignment equivalent.
It should only be used within other functions, where the expected data
structure can be guaranteed.
}
