% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.geolocatordp.R
\name{plot.geolocatordp}
\alias{plot.geolocatordp}
\title{Plot GeoLocator Data Package coverage}
\usage{
\method{plot}{geolocatordp}(x, type = NULL, ...)
}
\arguments{
\item{x}{a GeoLocator \code{pkg} object.}

\item{type}{type of the plot to display. One of \code{"coverage"}.}

\item{...}{additional parameters}
}
\value{
a plot, ggplotly or leaflet object.
}
\description{
Plot GeoLocator Data Package coverage
}
