% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressurepaths.R
\name{pressurepaths}
\alias{pressurepaths}
\alias{pressurepaths<-}
\title{Get or set pressurepaths}
\usage{
pressurepaths(x)

pressurepaths(x) <- value
}
\arguments{
\item{x}{GeoLocator Data Package object, as returned by \code{read_geolocatordp()}.}

\item{value}{A data frame to assign as pressurepaths}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}} data frame with pressurepaths
}
\description{
\code{pressurepaths()} gets the pressurepaths from a GeoLocator Data Package object.

\verb{pressurepaths<-()} is the assignment equivalent.
It should only be used within other functions, where the expected data
structure can be guaranteed.
}
