% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_soi.R
\name{read_soi}
\alias{read_soi}
\title{Read SOI data into a GeoLocator Data Package}
\usage{
read_soi(gdl, directory_data, generate_observations = TRUE)
}
\arguments{
\item{gdl}{A data frame containing the SOI data. Must include columns like \code{OrderName},
\code{GDL_ID}, and other relevant fields for tags, measurements, and observations. See \code{read_soi_gld()}
for more information.}

\item{directory_data}{A character string specifying the path to the directory where data files
are located. This directory is used to locate and match GDL_IDs to their corresponding
directories.}

\item{generate_observations}{A logical value indicating whether to create pre-filled
observations with missing values (date, locations, etc... assuming equipment and retrieval.}
}
\value{
A \code{geolocatordp} object.
}
\description{
This function reads data from the Swiss Ornithological Institute (SOI) into
a new package created with \code{\link[=create_gldp]{create_gldp()}}. It includes
tags, measurements, and observations based on the provided data frame and directory of data. The
function also handles missing directories and updates the package accordingly.

See an example of use \href{https://rpubs.com/GeoPressure/Geolocator_create_from_soi}{with this tutorial}.
}
\details{
The function performs the following steps:
\itemize{
\item Checks and retrieves the directory information for each GDL_ID/tag_id.
\item Creates GeoPressureR tag data for each of them when possible
\item Extract measurements and add them as resources to pkg
\item Compute tags.csv and observations.csv from \code{gdl} and add them as resources too.
}
}
