% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staps.R
\name{staps}
\alias{staps}
\alias{staps<-}
\title{Get or set staps}
\usage{
staps(x)

staps(x) <- value
}
\arguments{
\item{x}{GeoLocator Data Package object, as returned by \code{read_geolocatordp()}.}

\item{value}{A data frame to assign as staps}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}} data frame with staps
}
\description{
\code{staps()} gets the staps from a GeoLocator Data Package object.

\verb{staps<-()} is the assignment equivalent.
It should only be used within other functions, where the expected data
structure can be guaranteed.
}
