% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tags}
\alias{tags}
\alias{tags<-}
\title{Get or set tags}
\usage{
tags(x)

tags(x) <- value
}
\arguments{
\item{x}{GeoLocator Data Package object, as returned by \code{read_geolocatordp()}.}

\item{value}{A data frame to assign as tags}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}} data frame with tags
}
\description{
\code{tags()} gets the tags from a GeoLocator Data Package object.

\verb{tags<-()} is the assignment equivalent.
It should only be used within other functions, where the expected data
structure can be guaranteed.
}
