% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilights.R
\name{twilights}
\alias{twilights}
\alias{twilights<-}
\title{Get or set twilights}
\usage{
twilights(x)

twilights(x) <- value
}
\arguments{
\item{x}{GeoLocator Data Package object, as returned by \code{read_geolocatordp()}.}

\item{value}{A data frame to assign as twilights}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}} data frame with twilights
}
\description{
\code{twilights()} gets the twilights from a GeoLocator Data Package object.

\verb{twilights<-()} is the assignment equivalent.
It should only be used within other functions, where the expected data
structure can be guaranteed.
}
