% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecations.R, R/update_gldp.R
\name{update_gldp_metadata}
\alias{update_gldp_metadata}
\alias{update_gldp_spatial}
\alias{update_gldp_reference_location}
\alias{update_gldp}
\alias{update_gldp_order_resources}
\alias{update_gldp_temporal}
\alias{update_gldp_taxonomic}
\alias{update_gldp_number_tags}
\alias{update_gldp_bibliographic_citation}
\title{Update derived properties of a GeoLocator Data Package}
\usage{
update_gldp_metadata(pkg)

update_gldp_spatial(pkg)

update_gldp_reference_location(pkg)

update_gldp(pkg, ...)

update_gldp_order_resources(pkg)

update_gldp_temporal(pkg)

update_gldp_taxonomic(pkg)

update_gldp_number_tags(pkg)

update_gldp_bibliographic_citation(pkg, ...)
}
\arguments{
\item{pkg}{A GeoLocator Data Package object.}

\item{...}{Overrides passed to \code{\link[utils:bibentry]{utils::bibentry()}} when computing
\code{bibliographicCitation}.}
}
\value{
An updated GeoLocator Data Package object with derived properties
populated.
}
\description{
Updates and stores the following derived package properties:
\itemize{
\item \code{update_gldp_temporal()}: date range from \code{measurements$datetime}.
\item \code{update_gldp_taxonomic()}: species vector from \code{tags}, optionally filtered
to tags present in \code{measurements}.
\item \code{update_gldp_number_tags()}: counts by resource and sensor.
\item \code{update_gldp_bibliographic_citation()}: formatted citation from package
metadata.
\item \code{update_gldp_order_resources()}: reorder resources following schema order.
}

\code{update_gldp()} runs all functions above.
}
\seealso{
\code{\link[=read_gldp]{read_gldp()}} and \code{\link[=read_zenodo]{read_zenodo()}} call this function after reading,
and the resource setters such as \code{\link[=tags]{tags()}} and \code{\link[=measurements]{measurements()}} update
selected derived properties automatically.
}
