% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_metadata.R
\name{update_metadata}
\alias{update_metadata}
\title{Add computed metadata to a GeoLocator Data Package Metadata}
\usage{
update_metadata(pkg)
}
\arguments{
\item{pkg}{An object of class \code{"geolocatordp"} representing the GeoLocator Data Package to be
updated.}
}
\value{
An updated GeoLocator Data Package object of class \code{"geolocatordp"} with modified
metadata.
}
\description{
This function compute some metadata of an existing GeoLocator Data Package by setting or updating
several fields based on the data contained within the package. Specifically, it updates the
\code{created},\code{temporal}, \code{taxonomic}, and \code{reference_location} fields based on the data resources
present in the package.

The function performs the following updates:
\enumerate{
\item Sets the \code{created} field to the current date and time in RFC3339 format.
\item Updates the \code{temporal} field with the date range of the \code{datetime} values from the
\code{measurements} resource, if available.
\item Updates the \code{taxonomic} field with unique, non-missing \code{scientific_name} values from the
\code{tags} resource, if available.
\item Sets the \code{reference_location} field to the median latitude and longitude values from the
\code{observations} resource, if available.
}
}
