% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecations.R
\name{version}
\alias{version}
\title{Deprecated: \code{version()}}
\usage{
version(x)
}
\arguments{
\item{x}{A GeoLocator Data Package object.}
}
\value{
A GLDP version string.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{version()} was renamed to \code{\link[=gldp_version]{gldp_version()}}.
}
