% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo2gldp.R
\name{zenodo2gldp}
\alias{zenodo2gldp}
\title{Create a GeoLocator Data Package from Zenodo DOI}
\usage{
zenodo2gldp(
  doi,
  token = keyring::key_get(service = "ZENODO_PAT"),
  by = "byDOI"
)
}
\arguments{
\item{doi}{A string representing the DOI of the Zenodo Deposition to fetch metadata from (e.g.
\code{"10.5281/zenodo.14191023"}).}

\item{token}{A Zenodo API token. Best practice is to store it in your keyring (see description
above), in which case the default value will retrieve it using \code{keyring::key_get()}.}

\item{by}{The \verb{ZenodoManager$getDepositionBy*()} function used to retrieve the Deposition. Can
be either \code{"byDOI"}, \code{"ByConceptDOI"}, \code{"ById"}, or \code{"ByConceptId"}. Adjust the \code{doi} argument
accordingly.}
}
\value{
A Geolocator Data Package object
}
\description{
This function creates a \href{https://datapackage.org/standard/data-package/}{GeoLocator Data Package}
based on an existing Zenodo repository. It retrieves all metadata available using the Zenodo API
through the R package \href{https://github.com/eblondel/zen4R}{zen4R}.

To use this function, you'll need to:
\enumerate{
\item \href{https://help.zenodo.org/docs/get-started/create-an-account/}{Create a Zenodo account}
\item \href{https://zenodo.org/account/settings/applications/tokens/new/}{Generate an API token}
\item Store the token securely using \href{https://keyring.r-lib.org/}{\code{keyring}}:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{keyring::key_set_with_value(
  service = "ZENODO_PAT",
  password = "<your_api_token>"
)
}\if{html}{\out{</div>}}
}
}
\examples{
\dontrun{
zenodo2gldp("10.5281/zenodo.14191023")
}

}
