% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecations.R
\name{add_gldp_geopressuretemplate}
\alias{add_gldp_geopressuretemplate}
\title{Deprecated: \code{add_gldp_geopressuretemplate()}}
\usage{
add_gldp_geopressuretemplate(
  pkg,
  directory = ".",
  from = c("raw-tag", "interim")
)
}
\arguments{
\item{pkg}{A GeoLocator Data Package object.}

\item{directory}{A GeoPressureTemplate directory.}

\item{from}{Data sources to import.}
}
\value{
Updated package.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{add_gldp_geopressuretemplate()} was renamed to \code{\link[=read_geopressuretemplate]{read_geopressuretemplate()}}.
}
