% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gldp_geopressuretemplate.R
\name{add_gldp_geopressuretemplate}
\alias{add_gldp_geopressuretemplate}
\title{Add GLDP Geopressure Template}
\usage{
add_gldp_geopressuretemplate(pkg, directory, from = NULL, replace = FALSE)
}
\arguments{
\item{pkg}{A GLDP package object.}

\item{directory}{A character string specifying the directory where the data files are located.}

\item{from}{A character string specifying the source of the data files. Can be "raw-tag" (for
creating \code{tag} based on the data in \verb{data/raw-tag/}) or
"interim" for data in \code{data/interim}. If \code{NULL} (default), the function will determine the
source based on the presence of at least one file in the "interim" directory.}

\item{replace}{If \code{TRUE}, the added resource will replace an existing
resource with the same name.}
}
\value{
The updated GLDP package object with added geopressure templates.
}
\description{
This function adds geopressuretemplates to a GLDP package by reading data from specified
directories and files.
}
\details{
The function performs the following steps:
\enumerate{
\item Reads the "tags.csv" and "observations.csv" files from the "./data" directory if they exist
and adds them to the GLDP package.
\item Determines the source of the data files ("data" or "interim") if the \code{from} parameter is NULL.
\item Reads all ".RData" files from the specified source directory and processes them.
}

You can exclude interim file to be included in the package by starting the file name with an \verb{_}.

It is not possible to do a mix of some \code{tag} read from \code{from="data"} and some \code{tag} read from
\code{from="interim"}.
}
\examples{
\dontrun{
pkg <- create_gldp_package()
pkg <- add_gldp_geopressuretemplate(pkg, directory = "path/to/data")
}

}
