% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gldp_geopressuretemplate.R
\name{add_gldp_geopressuretemplate}
\alias{add_gldp_geopressuretemplate}
\title{Add Geolocator DP resources from a GeoPressureTemplate}
\usage{
add_gldp_geopressuretemplate(
  pkg,
  directory = ".",
  from = c("raw-tag", "interim")
)
}
\arguments{
\item{pkg}{A GeoLocator Data Package object.}

\item{directory}{A character string specifying the geopressuretemplate directory.}

\item{from}{A character vector specifying the source of the data files. Either or both of
\code{"raw-tag"} (for creating \code{tag} based on the data in \verb{data/raw-tag/}) and \code{"interim"} for data
in \code{data/interim}.}
}
\value{
The updated GLDP package object with new resources
}
\description{
This function adds all possible resources to a GeolocatoR Data Package by reading data from
a \href{https://github.com/Rafnuss/GeoPressureTemplate}{GeoPressureTemplate} directories and files.

The function performs the following steps:
\enumerate{
\item If \code{"interim"} in \code{from},  reads all interim ".RData" files, extract all variables and add
corresponding resources to the package (\code{measurements}, \code{twilights}, \code{staps}, \code{paths}, \code{edges},
and \code{pressurepaths}). Generated automatically temporary \code{tags} and \code{observations} tables from
the \code{param} in the interim file.
\item If \code{"raw-data"} in \code{from}, search for all tag_id in \verb{data/raw-data/} which were not included
from \code{interim}. It will use at least \code{tag_create} and if possible \code{tag_label} and \code{tag_set_map}
using \code{config.yml} .It will also generate or update \code{tags} and \code{observations} table from the the
tag data.
\item Reads the \code{tags.csv} (or \code{tags.xlsx} if present) and \code{observations.csv} (or
\code{observations.xlsx} if present) from the \code{./data} directory if they exist and overwrite the
previously generated \code{tags} and \code{observations}.
}

You can exclude interim file or raw-tag folder to be included in the package by starting the
file name with an \code{"_"}.

It is possible to do a mix of some \code{tag} read from \code{"interim"} and other from \code{"raw-data"}
simultaneously.

You can find more information on the use of this function in the \href{https://raphaelnussbaumer.com/GeoPressureManual/geolocator-create.html}{GeoPressureManual }
}
