% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gldp_geopressuretemplate.R
\name{add_gldp_geopressuretemplate}
\alias{add_gldp_geopressuretemplate}
\title{Add Geolocator DP resources from a GeoPressureTemplate}
\usage{
add_gldp_geopressuretemplate(
  pkg,
  directory = ".",
  from = c("raw-tag", "interim"),
  replace = FALSE
)
}
\arguments{
\item{pkg}{A GeoLocator Data Package object.}

\item{directory}{A character string specifying the geopressuretemplate directory.}

\item{from}{A character vector specifying the source of the data files. Either or both of
\code{"raw-tag"} (for creating \code{tag} based on the data in \verb{data/raw-tag/}) and \code{"interim"} for data
in \code{data/interim}.}

\item{replace}{If \code{TRUE}, the added resource will replace an existing
resource with the same name.}
}
\value{
The updated GLDP package object with new resources
}
\description{
This function adds all possible resources to a GeolocatoR Data Package by reading data from
a \href{https://github.com/Rafnuss/GeoPressureTemplate}{GeoPressureTemplate} directories and files.

The function performs the following steps:
\enumerate{
\item If \code{"interim"} in \code{from},  reads all interim ".RData" files, extract all resources possibles
and add them to the package. \code{tags} and \code{observations} are generated automatically from \code{param}.
\item If \code{"raw-data"} in \code{from}, create all the possible resources from \verb{data/raw-data/} and
\verb{data/tag-label/}. \code{tags} and \code{observations} resources are also generated from available
information.
\item Reads the \code{tags.csv} and \code{observations.csv} files from the \code{./data} directory if they exist
and overwrite the previously generated \code{tags} and \code{observations}.
}

You can exclude interim file to be included in the package by starting the file name with an \verb{_}.

It is possible to do a mix of some \code{tag} read from \code{"interim"} and other from \code{"raw-data"}
simultaneously.

You can find more information on the use of this function in the \href{https://raphaelnussbaumer.com/GeoPressureManual/geolocator-create.html}{GeoPressureManual }
}
