% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gldp_resource.R
\name{add_gldp_resource}
\alias{add_gldp_resource}
\title{Add a Geolocator Data Resource}
\usage{
add_gldp_resource(
  package,
  resource_name,
  data,
  cast_type = FALSE,
  replace = FALSE,
  delim = ","
)
}
\arguments{
\item{package}{A GeoLocator Data Package object to which the resource will be added.}

\item{resource_name}{A character string specifying the name of the resource. This name is used
to locate the schema file.}

\item{data}{A data frame containing the data to be added as a resource. The data frame will be
adjusted according to the schema.}

\item{cast_type}{A logical value indicating whether the data frame should be cast to the types
specified in the schema. Defaults to \code{FALSE}.}

\item{replace}{If \code{TRUE}, the added resource will replace an existing
resource with the same name.}

\item{delim}{Single character used to separate the fields in the CSV file(s),
e.g. \verb{\\t} for tab delimited file.
Will be set as \code{delimiter} in the resource Table Dialect, so read functions
.  know how to read the file(s).}
}
\value{
The updated GeoLocator Data Package object with the new resource added.
}
\description{
This function adds a resource to a geolocator data package and ensures that the data conforms to
the schema defined for that resource. It is a wrapper of \href{https://docs.ropensci.org/frictionless/reference/add_resource.html}{\code{frictionless::add_resource} } where it first validate
against the schema and potentially modify the data frame \code{data} before adding it to the package.

More specifically, the function adjusts the data frame according to the schema's
\href{https://datapackage.org/standard/table-schema/#fieldsMatch}{\code{fieldsMatch}} property and also
cast the type/class of the columns provided according to. \code{cast_type}.

Note that this function is generally not recommended to be used as all resources can be added or
modified with their respective \href{https://bit.ly/41HruRs}{accessors functions}.
}
