% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gldp_soi.R
\name{add_gldp_soi}
\alias{add_gldp_soi}
\title{Add GLDP SOI Data to a Package}
\usage{
add_gldp_soi(pkg, gdl, directory_data, generate_observations = TRUE)
}
\arguments{
\item{pkg}{The package object to which the data will be added.}

\item{gdl}{A data frame containing the SOI data. Must include columns like \code{OrderName},
\code{GDL_ID}, and other relevant fields for tags, measurements, and observations. See \code{\link{read_gdl}}
for more information.}

\item{directory_data}{A character string specifying the path to the directory where data files
are located. This directory is used to locate and match GDL_IDs to their corresponding
directories.}

\item{generate_observations}{A logical value indicating whether to create pre-filled
observations with missing values (date, locations, etc... assuming equipment and retrieval.}
}
\value{
The updated package object with the added resources.
}
\description{
This function adds data from the Swiss Ornithological Institute (SOI) to a package. It includes
tags, measurements, and observations based on the provided data frame and directory of data. The
function also handles missing directories and updates the package accordingly.

See an example of use \href{https://rpubs.com/rafnuss/geolocator_create_from_soi}{with this tutorial}.
}
\details{
The function performs the following steps:
\itemize{
\item Checks and retrieves the directory information for each GDL_ID/tag_id.
\item Creates GeoPressureR tag data for each of them when possible
\item Extract measurements and add them as resources to pkg
\item Compute tags.csv and observations.csv from \code{gdl} and add them as resources too.
}
}
