% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config2tibble.R
\name{config2tibble}
\alias{config2tibble}
\title{Convert config.yml to tibble
#' @param file Path to config.yml file. Default: "config.yml" or value of R_CONFIG_FILE environment variable.}
\usage{
config2tibble(
  file = Sys.getenv("R_CONFIG_FILE", "config.yml"),
  filter_return = TRUE
)
}
\value{
A tibble with one row per tag and columns for each configuration parameter.
}
\description{
Convert config.yml to tibble
#' @param file Path to config.yml file. Default: "config.yml" or value of R_CONFIG_FILE environment variable.
}
