% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_geopressuretemplate.R
\name{create_geopressuretemplate}
\alias{create_geopressuretemplate}
\title{Create a GeoPressureTemplate Project}
\usage{
create_geopressuretemplate(path, pkg = NULL, open = interactive())
}
\arguments{
\item{path}{A character string specifying the destination directory where the project will
be created. The last folder will give the name to the project.}

\item{pkg}{A GeoLocatoR Datapackage object (optional)}

\item{open}{If \code{TRUE}, the package is opened in a new RStudio session.}
}
\value{
The path to the created project directory.
}
\description{
Initializes a geopressure template project by creating a specified directory structure and
populating it with essential files, including a DESCRIPTION file, README, license, and data.

This function sets up the project directory and ensures that all necessary components are in
place for a geopressure analysis project.
}
\details{
Upon execution, the function performs the following steps:
\itemize{
\item Creates the project directory.
\item Generates the DESCRIPTION file using metadata from \code{pkg}.
\item Creates a README file that outlines project details.
\item Generates a LICENSE file based on the specified licenses.
\item Writes relevant data files into the project structure.
}
}
