% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gldp_geopressuretemplate.R
\name{create_gldp_geopressuretemplate}
\alias{create_gldp_geopressuretemplate}
\title{Create a GeoLocator Data Package from a GeoPressure Template}
\usage{
create_gldp_geopressuretemplate(directory = ".")
}
\arguments{
\item{directory}{A string specifying the path to the directory containing the
GeoPressureTemplate, including the \code{DESCRIPTION} file.}
}
\value{
A list containing the descriptor for the GeoLocator Data Package.
}
\description{
This function creates a GeoLocator Data Package based on the information found in a
\code{DESCRIPTION} file located in the specified directory. It utilizes the \code{create_gldp} function
and ensures that required fields are present, converting them into a format compatible with
GeoLocator Data Package standards.
}
\details{
The function performs the following steps:
\enumerate{
\item Checks if the provided directory exists.
\item Reads the \code{DESCRIPTION} file using the \code{desc} package.
\item Verifies the presence of mandatory fields (\code{Title}, \code{Authors@R}, and \code{License}) and warns
about the absence of suggested fields (\code{Version} and \code{Description}).
\item Maps author roles to their corresponding GeoLocator roles.
\item Creates and returns a GeoLocator Data Package with the information extracted from the
\code{DESCRIPTION} file.
}
}
