% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLocatoR-package.R
\docType{package}
\name{GeoLocatoR-package}
\alias{GeoLocatoR}
\alias{GeoLocatoR-package}
\title{GeoLocatoR: R package for GeoLocator Data Package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An R package to help you create and read geolocator data using the GeoLocator Data Package standard.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GeoPressure/Geolocator}
  \item \url{https://geopressure.org/GeoLocatoR/}
  \item Report bugs at \url{https://github.com/GeoPressure/GeoLocatoR/issues}
}

}
\author{
\strong{Maintainer}: Raphaël Nussbaumer \email{rafnuss@gmail.com} (\href{https://orcid.org/0000-0002-8185-1020}{ORCID})

Other contributors:
\itemize{
  \item Peter Desmet \email{peter.desmet@inbo.be} (\href{https://orcid.org/0000-0002-8442-8025}{ORCID}) [contributor]
}

}
\keyword{internal}
