% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLocatoR-package.R
\docType{package}
\name{GeoLocatoR-package}
\alias{GeoLocatoR}
\alias{GeoLocatoR-package}
\title{GeoLocatoR: R package for GeoLocator Data Package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

R package will help you create and read geolocator data using the GeoLocator Data package standard.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rafnuss/geolocator}
}

}
\author{
\strong{Maintainer}: Raphaël Nussbaumer \email{rafnuss@gmail.com} (\href{https://orcid.org/0000-0002-8185-1020}{ORCID})

}
\keyword{internal}
