% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gldp_to_eml.R
\name{gldp_to_eml}
\alias{gldp_to_eml}
\title{Transform a GeoLocator Data Package to EML}
\usage{
gldp_to_eml(package, directory)
}
\arguments{
\item{package}{A GeoLocator Data Package object.}

\item{directory}{Path to local directory to write files to.}
}
\value{
\code{eml.xml} file written to disk.
Invisibly, an \link[EML:eml]{EML::eml} object.
}
\description{
Transforms a \href{https://raphaelnussbaumer.com/GeoLocator-DP/}{GeoLocator Data Package}
to \href{https://eml.ecoinformatics.org/}{Ecological Metadata Language (EML)}.
The resulting EML file can be uploaded to an \href{https://www.gbif.org/ipt}{IPT} for
publication to GBIF and/or OBIS. A corresponding Darwin Core Archive can be created
with \code{\link[=gldp_to_dwc]{gldp_to_dwc()}}.
}
\section{Transformation details}{

Metadata are derived from the GeoLocator Data Package and transformed to EML.
The following properties are set:
\itemize{
\item \code{title}: Package title as provided in \code{package$title}.
\item \code{abstract}: Package description as provided in \code{package$description}.
\item \code{pubDate}: Publication year extracted from \code{package$created}.
\item \code{creator}: Contributors with roles \code{"ProjectLeader"}, \code{"Researcher"}, or
\code{"DataCurator"} as provided in \code{package$contributors}.
\item \code{contact}: Contributors with role \code{"ContactPerson"} as provided in
\code{package$contributors}.
\item \code{associatedParty}: Other contributors as provided in \code{package$contributors},
including those with role \code{"RightsHolder"}.
\item \code{intellectualRights}: License information from \code{package$licenses}.
\item \code{keywords}: Keywords as provided in \code{package$keywords}.
\item \code{packageId}: Package identifier as provided in \code{package$id}. If no ID is
provided, a UUID is generated. As a result, no new DOI will be created when
publishing to GBIF if \code{package$id} contains a DOI.
}

Coverage information:
\itemize{
\item \code{temporalCoverage}: Date range from \code{package$temporal} (if available),
derived from measurement timestamps.
\item \code{geographicCoverage}: Bounding box from \code{package$spatial} (if available),
calculated from all locations in observations, paths, and pressurepaths.
\item \code{taxonomicCoverage}: Species list from \code{package$taxonomic} (if available),
derived from unique scientific names in tags.
}

Methods:
\itemize{
\item \code{methodStep}: Generic description indicating data were processed using the
GeoLocator Data Package standard.
}

The following EML properties are not set:
\itemize{
\item \code{type}
\item \code{subtype}
\item \verb{update frequency}
\item \verb{publishing organization}
\item \verb{project data}
\item \code{citations}
\item \verb{collection data}
}
}

