% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versioning.R
\name{gldp_version}
\alias{gldp_version}
\title{Get GeoLocator DP version}
\usage{
gldp_version(x)
}
\arguments{
\item{x}{A GeoLocator Data Package object.}
}
\value{
Character string version (e.g. \code{"v1.0"}).

The function aborts if \verb{$schema} is missing, cannot be parsed, or refers to
an unsupported version.
}
\description{
Extract the GeoLocator-DP version from \code{x$schema}.
}
\concept{misc functions}
