% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_to_observations.R
\name{params_to_observations}
\alias{params_to_observations}
\title{Generate \code{observations} from GeoPressureR parameters}
\usage{
params_to_observations(params)
}
\arguments{
\item{params}{A list of GeoPressureR parameter objects. These parameters should have been
generated during the GeoPressure workflow. See \href{https://raphaelnussbaumer.com/GeoPressureR/reference/param_create.html}{\code{GeoPressureR::param_create()} } for more information.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} data frame with columns \code{ring_number}, \code{tag_id},
\code{observation_type}, \code{datetime}, \code{longitude}, \code{latitude}, and \code{comments}.
}
\description{
This function converts a list of parameters into a single \code{observations} tibble according to the
\href{https://raphaelnussbaumer.com/GeoLocator-DP/core/observations/}{GeoLocator Data Package specification }.
}
