% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.geolocatordp.R
\name{plot.geolocatordp}
\alias{plot.geolocatordp}
\title{Plot GeoLocator Data Package coverage}
\usage{
\method{plot}{geolocatordp}(x, type = NULL, ...)
}
\arguments{
\item{x}{A GeoLocator Data Package object.}

\item{type}{Type of the plot to display. Currently only \code{"coverage"} is supported.}

\item{...}{Additional parameters passed to plotting functions.}
}
\value{
a plot, ggplotly or leaflet object.
}
\description{
Creates visualizations for a GeoLocator Data Package, including coverage plots
showing data availability across time for different sensors and tags.
}
