% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.geolocatordp.R
\name{print.geolocatordp}
\alias{print.geolocatordp}
\title{Print a GeoLocator Data Package}
\usage{
\method{print}{geolocatordp}(x, ...)
}
\arguments{
\item{x}{A GeoLocator Data Package object, as returned by \code{read_gldp()}.}

\item{...}{Further arguments, they are ignored by this function.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints a human-readable summary of a GeoLocator Data Package, as an
extension of \code{\link[frictionless:print.datapackage]{print.datapackage()}}.
}
\examples{
pkg <- read_zenodo("17367319", quiet = TRUE)
print(pkg)

}
\concept{print functions}
