% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gldp.R
\name{read_gldp}
\alias{read_gldp}
\title{Read a GeoLocator Data Package}
\usage{
read_gldp(file = "datapackage.json")
}
\arguments{
\item{file}{A string specifying the path to the JSON file containing the GeoLocator Data Package
metadata. Defaults to \code{"datapackage.json"}.}
}
\value{
An object of class \code{"geolocatordp"} containing the metadata from the GeoLocator Data
Package.
}
\description{
This function reads a GeoLocator Data Package from a \code{"datapackage.json"}, file that describes
the Data Package metadata and its Data Resources. The function wraps the function
\href{https://docs.ropensci.org/frictionless/reference/read_package.html}{\code{frictionless::read_package} } and assigns the class
\code{"geolocatordp"} to the datapackage read.
}
\examples{
# Read a datapackage.json file
package <- read_gldp("https://zenodo.org/records/14099115/files/datapackage.json")

package

# Access the Data Package properties
package$name
package$created
}
