% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gldp.R
\name{read_gldp}
\alias{read_gldp}
\title{Read a GeoLocator Data Package}
\usage{
read_gldp(file = "datapackage.json")
}
\arguments{
\item{file}{A string specifying the path to the JSON file containing the GeoLocator Data Package
metadata. Defaults to \code{"datapackage.json"}.}
}
\value{
An object of class \code{"geolocatordp"} containing the metadata from the GeoLocator Data Package.
}
\description{
This function reads a GeoLocator Data Package from a specified JSON file. The file should conform
to the GeoLocator Data Package format. The function utilizes the \code{frictionless} package to read the
package metadata and then assigns the class \code{"geolocatordp"} to the resulting object for proper
identification and handling.
}
