% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_gldp.R
\name{update_gldp}
\alias{update_gldp}
\alias{update_gldp_temporal}
\alias{update_gldp_spatial}
\alias{update_gldp_taxonomic}
\alias{update_gldp_number_tags}
\alias{update_gldp_bibliographic_citation}
\alias{update_gldp_reference_location}
\alias{update_gldp_metadata}
\title{Update computed metadata of a GeoLocator Data Package}
\usage{
update_gldp(pkg)

update_gldp_temporal(pkg)

update_gldp_spatial(pkg)

update_gldp_taxonomic(pkg)

update_gldp_number_tags(pkg)

update_gldp_bibliographic_citation(pkg)

update_gldp_reference_location(pkg)

update_gldp_metadata(pkg)
}
\arguments{
\item{pkg}{A GeoLocator Data Package object}
}
\value{
An updated GeoLocator Data Package object with modified metadata.
}
\description{
This function update/create the following metadata of an existing GeoLocator Data Package
based on the data contained within the package.
\itemize{
\item \code{update_gldp_temporal()}: Sets \code{pkg$temporal} with the date range of the \code{datetime} values from
the \code{measurements} resource, if available.
\item \code{update_gldp_spatial()}: Sets \code{pkg$spatial} to the GeoJSON bounding box of all locations found
in observation, paths and pressurepaths.
\item \code{update_gldp_taxonomic()}: Sets \code{pkg$taxonomic} to the unique \code{tags$scientific_name} found in
the observations.
\item \code{update_gldp_number_tags()}:
\item \code{update_gldp_bibliographic_citation()}: Sets \code{pkg$bibliographicCitation} to the formatted
\code{utils::bibentry()} of the current package.
\item \code{update_gldp_reference_location()}: Sets the \code{pkg$reference_location} field to the median
latitude and longitude values from the \code{observations} resource, if available.
\item \code{update_gldp_metadata}: Re-arrange the order of the properties according to the order of the
schema
}

\code{update_gldp} performs all of the the functions mentioned above.
}
