% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_gldp.R
\name{validate_gldp}
\alias{validate_gldp}
\title{Validate a GeoLocator Data Package}
\usage{
validate_gldp(pkg, quiet = FALSE)
}
\arguments{
\item{pkg}{A GeoLocator Data Package object to be validated.}

\item{quiet}{A logical indicating whether to suppress messages from the \code{cli} package. Defaults
to \code{FALSE}.}
}
\value{
A logical value indicating whether the package validation was successful (\code{TRUE}) or
failed (\code{FALSE}).
}
\description{
This function performs a comprehensive validation of a GeoLocator Data Package by checking the
package metadata, profile, and resources. The validation includes verifying that the package
conforms to the GeoLocator Data Package profile and that each resource adheres to its schema.

If \code{quiet} is \code{TRUE}, the function suppresses the output of the \code{cli} package's messages.
}
