% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecations.R
\name{version}
\alias{version}
\title{Deprecated: \code{version()}}
\usage{
version(x)
}
\arguments{
\item{x}{A GeoLocator Data Package object.}
}
\value{
A GLDP version string.
}
\description{
\code{version()} was renamed to \code{\link[=gldp_version]{gldp_version()}}.
}
