% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{version}
\alias{version}
\title{Get GeoLocator DP version}
\usage{
version(x)
}
\arguments{
\item{x}{A GeoLocator Data Package object, as returned by
\code{create_gldp()}. Also works on a Frictionless Data Package, as returned by
\code{read_package()}.}
}
\value{
Character string with the GeoLocator DP version number (e.g. \code{"v0.2"}).
}
\description{
Extracts the version number used by a GeoLocator Data Package object.
This version number indicates what version of the \href{https://raphaelnussbaumer.com/GeoLocator-DP/}{GeoLocator DP standard} was used.

The version number is derived as follows:
\enumerate{
\item The \code{version} attribute, if defined.
\item A version number contained in \code{x$profile}, which is expected to
contain the URL to the used GeoLocator DP standard.
\item \code{x$profile} in its entirety (can be \code{NULL}).
}
}
\concept{misc functions}
