% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo_to_gldp.R
\name{zenodo_to_gldp}
\alias{zenodo_to_gldp}
\title{Create or update the metadata of a GeoLocator Data Package based on a Zenodo Record}
\usage{
zenodo_to_gldp(zenodo_record, pkg = NULL)
}
\arguments{
\item{zenodo_record}{A \code{zen4R} ZenodoRecord object from which to extract metadata.}

\item{pkg}{(optionall) A Geolocator Data Package object on which the metadata should be added. If
not provided, create a new one.}
}
\value{
An updated or new Geolocator Data Package object with metadata from the zenodo record.
}
\description{
This function update the metadata of a \href{https://datapackage.org/standard/data-package/}{GeoLocator Data Package } based on an \code{zen4R} ZenodoRecord provided. If
no package is provided (\code{pkg = NULL}), a new data package is created.

All metadata are retrieved with the Zenodo API through the R package
\href{https://github.com/eblondel/zen4R}{zen4R}.

Note that no data is actually retrieved/copied.
}
